+++ brogue-hd/brogueRememberedCell.h
@@ -8,8 +8,8 @@ namespace brogueHd::backend::model
 	{
 		brogueCellDisplay rememberedAppearance;			// how the player remembers the cell to look
 		enum itemCategory rememberedItemCategory;		// what category of item the player remembers lying there
-		short rememberedItemKind;                       // what kind of item the player remembers lying there
-		short rememberedItemQuantity;                   // how many of the item the player remembers lying there
+		int rememberedItemKind;                       // what kind of item the player remembers lying there
+		int rememberedItemQuantity;                   // how many of the item the player remembers lying there
 		enum tileType rememberedTerrain;				// what the player remembers as the terrain (i.e. highest priority terrain upon last seeing)
 		unsigned long rememberedCellFlags;              // map cell flags the player remembers from that spot
 		unsigned long rememberedTerrainFlags;           // terrain flags the player remembers from that spot
